backend/apps/core/pagination.py
"""Custom pagination classes."""
from rest_framework.pagination import PageNumberPagination


class StandardPagination(PageNumberPagination):
    """Standard pagination with configurable page size."""
    
    page_size = 20
    page_size_query_param = "page_size"
    max_page_size = 100


class SmallPagination(PageNumberPagination):
    """Smaller pagination for nested resources like comments."""
    
    page_size = 10
    page_size_query_param = "page_size"
    max_page_size = 50
