backend/requirements/base.txt
# Django 6.0
Django>=6.0,<7.0

# Django REST Framework
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.28.0

# Database
psycopg[binary,pool]>=3.2.0

# Redis
redis>=5.0.0
django-redis>=5.4.0

# Async
uvicorn[standard]>=0.32.0

# Environment
python-dotenv>=1.0.0

# UUIDv7
uuid7>=0.1.0

# CORS
django-cors-headers>=4.3.0

# Filtering & Pagination
django-filter>=24.0

# Phone number field
django-phonenumber-field[phonenumbers]>=8.0.0
